Add camera movement system

+++ src/Components/BoxColliderComponent.h
@@ -4,11 +4,11 @@
 #include <glm/glm.hpp>
 
 struct BoxColliderComponent {
-    int width;
-    int height;
+    float width;
+    float height;
     glm::vec2 offset;
 
-    BoxColliderComponent(int width = 0, int height = 0, glm::vec2 offset = glm::vec2(0)) {
+    BoxColliderComponent(float width = 0, float height = 0, glm::vec2 offset = glm::vec2(0)) {
         this->width = width;
         this->height = height;
         this->offset = offset;
